# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Finance Advisor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wealth/snapshots:
    get:
      operationId: getWealthSnapshots
      tags: [wealth]
      summary: Get wealth management monthly snapshots
      responses:
        "200":
          description: Wealth snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WealthSnapshot"
    post:
      operationId: createWealthSnapshot
      tags: [wealth]
      summary: Create a wealth snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWealthSnapshotBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WealthSnapshot"
  /wealth/snapshots/{id}:
    patch:
      operationId: updateWealthSnapshot
      tags: [wealth]
      summary: Update a wealth snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWealthSnapshotBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WealthSnapshot"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteWealthSnapshot
      tags: [wealth]
      summary: Delete a wealth snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /wealth/credit-scores:
    get:
      operationId: getCreditScores
      tags: [wealth]
      summary: Get credit score history
      responses:
        "200":
          description: Credit scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditScore"
    post:
      operationId: createCreditScore
      tags: [wealth]
      summary: Record credit scores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditScoreBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditScore"

  # ─── Debt ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Assumptions ─────────────────────────────────────────────────────────
...
    WealthSnapshot:
      type: object
      properties:
        id:
          type: integer
        snapshotDate:
          type: string
          format: date
        hysa:
          type: number
        brokerage:
          type: number
        retirement401k:
          type: number
        otherAssets:
          type: number
        totalAssets:
          type: number
        carLoan:
          type: number
        studentLoans:
          type: number
        otherLiabilities:
          type: number
        totalLiabilities:
          type: number
        netWorth:
          type: number
        changeVsPrior:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - snapshotDate
        - hysa
        - brokerage
        - retirement401k
        - otherAssets
        - totalAssets
        - carLoan
        - studentLoans
        - otherLiabilities
        - totalLiabilities
        - netWorth
    CreateWealthSnapshotBody:
      type: object
      properties:
        snapshotDate:
          type: string
          format: date
        hysa:
          type: number
        brokerage:
          type: number
        retirement401k:
          type: number
        otherAssets:
          type: number
        carLoan:
          type: number
        studentLoans:
          type: number
        otherLiabilities:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - snapshotDate
        - hysa
        - brokerage
        - retirement401k
        - otherAssets
        - carLoan
        - studentLoans
        - otherLiabilities
    UpdateWealthSnapshotBody:
      type: object
      properties:
        snapshotDate:
          type: string
          format: date
        hysa:
          type: number
        brokerage:
          type: number
        retirement401k:
          type: number
        otherAssets:
          type: number
        carLoan:
          type: number
        studentLoans:
          type: number
        otherLiabilities:
          type: number
        notes:
          type: ["string", "null"]
    CreditScore:
      type: object
      properties:
        id:
          type: integer
        asOfDate:
          type: string
          format: date
        experian:
          type: ["integer", "null"]
        equifax:
          type: ["integer", "null"]
        transunion:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - asOfDate
    CreateCreditScoreBody:
      type: object
      properties:
        asOfDate:
          type: string
          format: date
        experian:
          type: ["integer", "null"]
        equifax:
          type: ["integer", "null"]
        transunion:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - asOfDate

    # ─── Debt ─────────────────────────────────────────────────────────────────
...