# lib/api-zod/src/generated/types/creditScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */

export interface CreditScore {
  id: number;
  asOfDate: Date;
  /** @nullable */
  experian?: number | null;
  /** @nullable */
  equifax?: number | null;
  /** @nullable */
  transunion?: number | null;
  /** @nullable */
  notes?: string | null;
}